request location service when search page open for iOS
